+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add configurable FEC redundancy [#82]
 - Add configurable UDP send interval [#83]
 - Add UDP network tweak configuration [#86]
+- Add logs to pending RwLock [#92]
 
 ### Fixed
 
@@ -64,3 +65,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#82]: https://github.com/dusk-network/kadcast/issues/82
 [#83]: https://github.com/dusk-network/kadcast/issues/83
 [#90]: https://github.com/dusk-network/kadcast/issues/90
+[#92]: https://github.com/dusk-network/kadcast/issues/92